requirements.txt
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.8
defusedxml==0.7.1
dominate==2.9.1
et_xmlfile==2.0.0
fastjsonschema==2.21.1
Flask==3.1.0
flask-bootstrap5==0.1.dev1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
mistune==3.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
platformdirs==4.3.6
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.36.1
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
